requirements.txt
Django>=4.0.2,<4.1
celery[redis]==5.2.2,<5.3
django-crispy-forms
django-extensions
factory_boy
psycopg2>=2.8.6,<2.9
pdbpp>=0.10.3,<0.11.0
# selenium
openpyxl==2.6.4
# Pillow
# setproctitle
uWSGI>=2.0.19.1,<2.1
djangorestframework>=3.13.1,<3.14
